fix warnings

+++ src/main.cpp
@@ -14,9 +14,10 @@
 #include <unistd.h>
 #include <syslog.h>
 using namespace std;
+using namespace sysuh3c;
 
 int lockfd = -1;
-static const char *lockfname = "/tmp/sysuh3c.lock";
+static char *lockfname = nullptr;
 
 static void interrupt_handler(int signo) {
     if (lockf(lockfd, F_ULOCK, 0) < 0) exit(EXIT_FAILURE);
@@ -40,7 +41,18 @@ void daemonize() {
     close(fileno(stderr));
 
     umask(027);
-    chdir("/");
+    if (chdir("/") == -1) {
+        perror("chdir");
+        abort();
+    }
+
+    // Generate a tempfile name
+    char tmpfname_template[] = "sysuh3c_XXXXXX";
+    lockfname = mktemp(tmpfname_template);
+    if (lockfname) {
+        cerr << "Fail to generate a tempfile name" << endl;
+        abort();
+    }
 
     lockfd = open(lockfname, O_RDWR | O_CREAT, 0640);
     if (lockfd < 0) exit(EXIT_FAILURE);
@@ -48,7 +60,10 @@ void daemonize() {
 
     char pidstr[128] = {0};
     sprintf(pidstr, "%d\n", getpid());
-    write(lockfd, pidstr, strlen(pidstr));
+    if (write(lockfd, pidstr, strlen(pidstr)) == -1) {
+        perror("write");
+        abort();
+    }
 
     openlog("sysuh3c", LOG_CONS, LOG_USER);
 
@@ -62,7 +77,7 @@ void daemonize() {
     is_daemon = true;
 }
 
-int main(int argc, char **argv) {
+int main(int argc, char *const argv[]) {
 
     if (geteuid() != 0) {
         cerr << "You have to run the program as root" << endl;
@@ -123,7 +138,7 @@ int main(int argc, char **argv) {
         char *pwd = getpass("Password: ");
         password = pwd;
     }
-    
+
     EAPAuth authservice(name, password, iface);
 
     authservice.set_promote_listener([] (const string& msg) {
@@ -170,7 +185,10 @@ int main(int argc, char **argv) {
     FILE *fp = fopen(lockfname, "r");
     if (fp != nullptr) {
         pid_t pid;
-        fscanf(fp, "%d", &pid);
+        if (fscanf(fp, "%d", &pid) == EOF) {
+            perror("fscanf");
+            abort();
+        }
         int ret = kill(pid, SIGINT);
         fclose(fp);
     }
